🐛Fix : deploy 문제제
